added zuul cors fix + added requestbody on weathercontroller
